Tests/LibWeb/Text/expected/SVG/use-element-does-not-reclone-on-load.txt

Had resolved instance root before load: true
Had forward instance root before load: false
Kept resolved instance root after load: true
Has forward instance root after load: true
